Fixed exception occuring in redraw() if the canvas was empty.
If the canvas was empty, calling .bbox('all') on it will return the None type. This caused an unhandled exception since we subsequently create a list from this None (which was expected to be a tuple). The function now handels that exception (by discarding it, no need to center anything if the canvas is empty).