point_cloud: Clear canvas, multiple sets (color)
Instead of overwriting the previous set, the option 'New uniform' now creates an independent sample. For this purpose, the 'points' key in the 'data' dictionary now points towards a list that contains (point_array, color) tuples. In the future, the user will have the choice to choose a color for each new data set.

Also, there is now a button to clear any points from the canvas.