Enabled scrolling the canvas with mouse wheel
This enables scrolling, but in a rudimentary fashion.
First, event.delta seems to be always 0 on Linux, but should rather give an indication of scroll distance. Thus, a constant value of '1' was used instead.
Second, this will always scroll the canvas, regardles of where the cursor is. Ideally, it should only scroll the canvas if the cursor hovers over it (and scroll something else otherwise).