Finalized plugin unload support
Plugins can now be unloaded by providing an unload() method. This will be called when the plugin is unloaded and should be used to clean up everything the plugin did. A plugin that was once loaded can be loaded again just by its name, the path is not needed. To facilitate this, the Plugin_Manager now stores plugins in a dictionary, with plugin names as keys, and the value being a 2-tuple of (plugin_object, path), where plugin_object is None if the plugin is not loaded.

All of this is used by the new 'Plugin' menu in the menubar of gui.py. It shows a checklist of every plugin that was loaded and can be used to load and unload plugins.